Moved ATMLib2 and songs to their own dirs
