+++ evade/sound/song.h
@@ -1,7 +1,8 @@
 #ifndef SONG_H
 #define SONG_H
 
-#include "atm_cmd_constants.h"
+//TODO: use <atm_cmd_constants.h> once we publish ATMLib2
+#include "../src/ATMLib2/atm_cmd_constants.h"
 
 #define ARRAY_SIZE(a) (sizeof (a) / sizeof ((a)[0]))
 #define NUM_PATTERNS(struct_) (ARRAY_SIZE( ((struct_ *)0)->patterns_offset))
