+++ evade/src/ATMLib2/osc.h
